Added contex and new struct ContextBody
